+++ lib/src/pluto_dual_grid.dart
@@ -1,49 +1,32 @@
 part of '../pluto_grid.dart';
 
-typedef PlutoDualOnSelectedEventCallback = void Function(
-    PlutoDualOnSelectedEvent event);
-
-class PlutoDualOnSelectedEvent {
-  PlutoOnSelectedEvent gridA;
-  PlutoOnSelectedEvent gridB;
-
-  PlutoDualOnSelectedEvent({
-    this.gridA,
-    this.gridB,
-  });
-}
-
-class PlutoDualGridProps {
-  final List<PlutoColumn> columns;
-  final List<PlutoRow> rows;
-  final PlutoOnLoadedEventCallback onLoaded;
-  final PlutoOnChangedEventCallback onChanged;
-  final CreateHeaderCallBack createHeader;
-  final CreateFooterCallBack createFooter;
-  final PlutoConfiguration configuration;
-
-  PlutoDualGridProps({
-    this.columns,
-    this.rows,
-    this.onLoaded,
-    this.onChanged,
-    this.createHeader,
-    this.createFooter,
-    this.configuration,
-  });
-}
-
 class PlutoDualGrid extends StatefulWidget {
   final PlutoDualGridProps gridPropsA;
+
   final PlutoDualGridProps gridPropsB;
+
   final PlutoMode mode;
+
   final PlutoDualOnSelectedEventCallback onSelected;
+
+  /// [PlutoDualGridDisplayRatio]
+  /// Set the width of the two grids by specifying the ratio of the left grid.
+  /// 0.5 is 5(left grid):5(right grid).
+  /// 0.8 is 8(left grid):2(right grid).
+  ///
+  /// [PlutoDualGridDisplayFixedAndExpanded]
+  /// Fix the width of the left grid.
+  ///
+  /// [PlutoDualGridDisplayExpandedAndFixed]
+  /// Fix the width of the right grid.
+  final PlutoDualGridDisplay display;
 
   PlutoDualGrid({
     this.gridPropsA,
     this.gridPropsB,
     this.mode,
     this.onSelected,
+    this.display = const PlutoDualGridDisplayRatio(),
   });
 
   @override
@@ -58,11 +41,11 @@ class _PlutoDualGridState extends State<PlutoDualGrid> {
   Widget _buildGrid({
     PlutoDualGridProps props,
     PlutoMode mode,
-    BoxConstraints size,
+    double width,
     bool isGridA,
   }) {
     return SizedBox(
-      width: size.maxWidth / 2,
+      width: width,
       child: PlutoGrid.popup(
         columns: props.columns,
         rows: props.rows,
@@ -130,17 +113,92 @@ class _PlutoDualGridState extends State<PlutoDualGrid> {
           _buildGrid(
             props: widget.gridPropsA,
             mode: widget.mode,
-            size: size,
+            width: widget.display.gridAWidth(size),
             isGridA: true,
           ),
           _buildGrid(
             props: widget.gridPropsB,
             mode: widget.mode,
-            size: size,
+            width: widget.display.gridBWidth(size),
             isGridA: false,
           ),
         ],
       );
     });
   }
+}
+
+typedef PlutoDualOnSelectedEventCallback = void Function(
+    PlutoDualOnSelectedEvent event);
+
+class PlutoDualOnSelectedEvent {
+  PlutoOnSelectedEvent gridA;
+  PlutoOnSelectedEvent gridB;
+
+  PlutoDualOnSelectedEvent({
+    this.gridA,
+    this.gridB,
+  });
+}
+
+abstract class PlutoDualGridDisplay {
+  double gridAWidth(BoxConstraints size);
+
+  double gridBWidth(BoxConstraints size);
+}
+
+class PlutoDualGridDisplayRatio implements PlutoDualGridDisplay {
+  final double ratio;
+
+  const PlutoDualGridDisplayRatio({
+    this.ratio = 0.5,
+  }) : assert(0 < ratio && ratio < 1);
+
+  double gridAWidth(BoxConstraints size) => size.maxWidth * ratio;
+
+  double gridBWidth(BoxConstraints size) => size.maxWidth * (1 - ratio);
+}
+
+class PlutoDualGridDisplayFixedAndExpanded implements PlutoDualGridDisplay {
+  final double width;
+
+  const PlutoDualGridDisplayFixedAndExpanded({
+    this.width = 206.0,
+  });
+
+  double gridAWidth(BoxConstraints size) => width;
+
+  double gridBWidth(BoxConstraints size) => size.maxWidth - width;
+}
+
+class PlutoDualGridDisplayExpandedAndFixed implements PlutoDualGridDisplay {
+  final double width;
+
+  const PlutoDualGridDisplayExpandedAndFixed({
+    this.width = 206.0,
+  });
+
+  double gridAWidth(BoxConstraints size) => size.maxWidth - width;
+
+  double gridBWidth(BoxConstraints size) => width;
+}
+
+class PlutoDualGridProps {
+  final List<PlutoColumn> columns;
+  final List<PlutoRow> rows;
+  final PlutoOnLoadedEventCallback onLoaded;
+  final PlutoOnChangedEventCallback onChanged;
+  final CreateHeaderCallBack createHeader;
+  final CreateFooterCallBack createFooter;
+  final PlutoConfiguration configuration;
+
+  PlutoDualGridProps({
+    this.columns,
+    this.rows,
+    this.onLoaded,
+    this.onChanged,
+    this.createHeader,
+    this.createFooter,
+    this.configuration,
+  });
 }