+++ test/src/manager/pluto_key_manager_test.dart
@@ -4,6 +4,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:pluto_grid/pluto_grid.dart';
 
+import '../../helper/pluto_widget_test_helper.dart';
 import '../../mock/mock_pluto_state_manager.dart';
 
 void main() {
@@ -285,4 +286,201 @@ void main() {
       ]));
     },
   );
+
+  group('_handleHomeEnd', () {
+    final withKeyboardListener =
+        PlutoWidgetTestHelper('키 입력 테스트', (tester) async {
+      final PlutoKeyManager keyManager = PlutoKeyManager(
+        stateManager: stateManager,
+      );
+
+      keyManager.init();
+
+      await tester.pumpWidget(
+        MaterialApp(
+          home: Material(
+            child: RawKeyboardListener(
+              onKey: (event) {
+                keyManager.subject.add(KeyManagerEvent(
+                  focusNode: FocusNode(),
+                  event: event,
+                ));
+              },
+              focusNode: keyboardFocusNode,
+              child: TextField(),
+            ),
+          ),
+        ),
+      );
+
+      // when
+      keyboardFocusNode.requestFocus();
+
+      await tester.pumpAndSettle();
+    });
+
+    withKeyboardListener.test('home', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.home);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.home);
+
+      // then
+      verify(stateManager.moveCurrentCellToEdgeOfColumns(MoveDirection.Left))
+          .called(1);
+    });
+
+    withKeyboardListener.test('home + shift', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.shift);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.home);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.home);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.shift);
+
+      // then
+      verify(stateManager.moveSelectingCellToEdgeOfColumns(MoveDirection.Left))
+          .called(1);
+    });
+
+    withKeyboardListener.test('home + ctrl', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.control);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.home);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.home);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.control);
+
+      // then
+      verify(stateManager.moveCurrentCellToEdgeOfRows(MoveDirection.Up))
+          .called(1);
+    });
+
+    withKeyboardListener.test('home + ctrl + shift', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.shift);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.control);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.home);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.home);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.control);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.shift);
+
+      // then
+      verify(stateManager.moveSelectingCellToEdgeOfRows(MoveDirection.Up))
+          .called(1);
+    });
+
+    withKeyboardListener.test('end', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.end);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.end);
+
+      // then
+      verify(stateManager.moveCurrentCellToEdgeOfColumns(MoveDirection.Right))
+          .called(1);
+    });
+
+    withKeyboardListener.test('end + shift', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.shift);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.end);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.end);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.shift);
+
+      // then
+      verify(stateManager.moveSelectingCellToEdgeOfColumns(MoveDirection.Right))
+          .called(1);
+    });
+
+    withKeyboardListener.test('end + ctrl', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.control);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.end);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.end);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.control);
+
+      // then
+      verify(stateManager.moveCurrentCellToEdgeOfRows(MoveDirection.Down))
+          .called(1);
+    });
+
+    withKeyboardListener.test('end + ctrl + shift', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.shift);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.control);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.end);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.end);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.control);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.shift);
+
+      // then
+      verify(stateManager.moveSelectingCellToEdgeOfRows(MoveDirection.Down))
+          .called(1);
+    });
+  });
+
+  group('_handlePageUpDown', () {
+    final withKeyboardListener =
+        PlutoWidgetTestHelper('키 입력 테스트', (tester) async {
+      final PlutoKeyManager keyManager = PlutoKeyManager(
+        stateManager: stateManager,
+      );
+
+      keyManager.init();
+
+      when(stateManager.offsetHeight).thenReturn(230);
+      when(stateManager.currentRowIdx).thenReturn(0);
+
+      await tester.pumpWidget(
+        MaterialApp(
+          home: Material(
+            child: RawKeyboardListener(
+              onKey: (event) {
+                keyManager.subject.add(KeyManagerEvent(
+                  focusNode: FocusNode(),
+                  event: event,
+                ));
+              },
+              focusNode: keyboardFocusNode,
+              child: TextField(),
+            ),
+          ),
+        ),
+      );
+
+      // when
+      keyboardFocusNode.requestFocus();
+
+      await tester.pumpAndSettle();
+    });
+
+    withKeyboardListener.test('pageUp', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.pageUp);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.pageUp);
+
+      // then
+      verify(stateManager.moveCurrentCellByRowIdx(-5, MoveDirection.Up))
+          .called(1);
+    });
+
+    withKeyboardListener.test('pageUp + shift', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.shift);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.pageUp);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.pageUp);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.shift);
+
+      // then
+      verify(stateManager.moveSelectingCellByRowIdx(-5, MoveDirection.Up))
+          .called(1);
+    });
+
+    withKeyboardListener.test('pageDown', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.pageDown);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.pageDown);
+
+      // then
+      verify(stateManager.moveCurrentCellByRowIdx(5, MoveDirection.Down))
+          .called(1);
+    });
+
+    withKeyboardListener.test('pageDown + shift', (tester) async {
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.shift);
+      await tester.sendKeyDownEvent(LogicalKeyboardKey.pageDown);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.pageDown);
+      await tester.sendKeyUpEvent(LogicalKeyboardKey.shift);
+
+      // then
+      verify(stateManager.moveSelectingCellByRowIdx(5, MoveDirection.Down))
+          .called(1);
+    });
+  });
 }